+++ src/main/java/com/contactlab/api/ws/SendImmediateByCampaignToSubscriberId.java
@@ -37,7 +37,7 @@ import com.contactlab.api.ws.domain.SendImmediateOptions;
  *         &lt;element name="userKey" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
  *         &lt;element name="campaign" type="{domain.ws.api.contactlab.com}Campaign" minOccurs="0"/>
  *         &lt;element name="sourceId" type="{http://www.w3.org/2001/XMLSchema}int"/>
- *         &lt;element name="subscriberId" type="{http://www.w3.org/2001/XMLSchema}int"/>
+ *         &lt;element name="subscriberId" type="{http://www.w3.org/2001/XMLSchema}long"/>
  *         &lt;element name="sendImmediateOptions" type="{domain.ws.api.contactlab.com}SendImmediateOptions" minOccurs="0"/>
  *       &lt;/sequence>
  *     &lt;/restriction>
@@ -62,7 +62,7 @@ public class SendImmediateByCampaignToSubscriberId {
     protected String userKey;
     protected Campaign campaign;
     protected int sourceId;
-    protected int subscriberId;
+    protected long subscriberId;
     protected SendImmediateOptions sendImmediateOptions;
 
     /**
@@ -157,7 +157,7 @@ public class SendImmediateByCampaignToSubscriberId {
      * Gets the value of the subscriberId property.
      * 
      */
-    public int getSubscriberId() {
+    public long getSubscriberId() {
         return subscriberId;
     }
 
@@ -165,7 +165,7 @@ public class SendImmediateByCampaignToSubscriberId {
      * Sets the value of the subscriberId property.
      * 
      */
-    public void setSubscriberId(int value) {
+    public void setSubscriberId(long value) {
         this.subscriberId = value;
     }
