+++ src/main/java/com/contactlab/api/ws/SendImmediateMessageSidCAlCA.java
@@ -38,7 +38,7 @@ import com.contactlab.api.ws.domain.Attachment;
  *         &lt;element name="userKey" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
  *         &lt;element name="campaignAlias" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
  *         &lt;element name="sourceIdentifier" type="{http://www.w3.org/2001/XMLSchema}int"/>
- *         &lt;element name="subscriberIdentifier" type="{http://www.w3.org/2001/XMLSchema}int"/>
+ *         &lt;element name="subscriberIdentifier" type="{http://www.w3.org/2001/XMLSchema}long"/>
  *         &lt;element name="includeDefaultAttachments" type="{http://www.w3.org/2001/XMLSchema}boolean"/>
  *         &lt;element name="attachments" type="{domain.ws.api.contactlab.com}Attachment" maxOccurs="unbounded" minOccurs="0"/>
  *       &lt;/sequence>
@@ -65,7 +65,7 @@ public class SendImmediateMessageSidCAlCA {
     protected String userKey;
     protected String campaignAlias;
     protected int sourceIdentifier;
-    protected int subscriberIdentifier;
+    protected long subscriberIdentifier;
     protected boolean includeDefaultAttachments;
     protected List<Attachment> attachments;
 
@@ -161,7 +161,7 @@ public class SendImmediateMessageSidCAlCA {
      * Gets the value of the subscriberIdentifier property.
      * 
      */
-    public int getSubscriberIdentifier() {
+    public long getSubscriberIdentifier() {
         return subscriberIdentifier;
     }
 
@@ -169,7 +169,7 @@ public class SendImmediateMessageSidCAlCA {
      * Sets the value of the subscriberIdentifier property.
      * 
      */
-    public void setSubscriberIdentifier(int value) {
+    public void setSubscriberIdentifier(long value) {
         this.subscriberIdentifier = value;
     }
